Fixed Warnings

+++ LambdaEngine/Source/Networking/API/NetworkDebugger.cpp
@@ -143,7 +143,7 @@ namespace LambdaEngine
 					ImGui::Text("%d", pSegmentPool->GetFreeSegments());
 					ImGui::Text("%.1f ms", pStatistics->GetPing().AsMilliSeconds());
 
-					pair.second.PingValues[s_PingValuesOffset] = pStatistics->GetPing().AsMilliSeconds();
+					pair.second.PingValues[s_PingValuesOffset] = (float)pStatistics->GetPing().AsMilliSeconds();
 					ImGui::PlotLines("", pair.second.PingValues.data(), pair.second.PingValues.size(), s_PingValuesOffset, "", 0.0f, 50.0f, ImVec2(0, 80.0f));
 				}
